requirements.txt
pip~=25.1.1
python-box~=7.3.2
attrs~=25.1.0
distro~=1.9.0
wheel~=0.41.2
protobuf~=5.29.3
pillow~=11.1.0
tornado~=6.4.2
Jinja2~=3.1.5
setuptools~=80.9.0
streamlit~=1.43.0
requests~=2.32.3
python-dotenv~=1.1.1
